Everything up to date. Still can't get the GET request to properly function
