+++ auth/aunthenticate.js
@@ -1,10 +1,12 @@
 const jwt = require('jsonwebtoken');
 
-const jwtKey = require('../_secrets/keys').jwtKey;
+const jwtKey = require('../_secrets/keys').jwtKey
+
 
 // quickly see what this file exports
 module.exports = {
   authenticate,
+  tokenGenerator
 };
 
 // implementation details
@@ -25,3 +27,13 @@ function authenticate(req, res, next) {
     });
   }
 }
+function tokenGenerator(user){
+    const payload ={
+      username: user.username
+    }
+    const options = {
+      expiresIn : '1h'
+    }
+    return jwt.sign(payload,jwtKey,options)
+  }
+
