+++ package.json
@@ -22,7 +22,7 @@
     "bcryptjs": "^2.4.3",
     "cors": "^2.8.4",
     "express": "^4.16.4",
-    "jsonwebtoken": "^8.3.0",
+    "jsonwebtoken": "^8.4.0",
     "knex": "^0.15.2",
     "sqlite3": "^4.0.2"
   },
